third_party/blink/web_tests/platform/generic/external/wpt/html/semantics/embedded-content/media-elements/ready-states/autoplay-hidden.optional-expected.txt

This is a testharness.js-based test.
FAIL Allow delaying autoplay until video elements become visible assert_equals: Expected pause event, but got playing event instead expected "pause" but got "playing"
Harness: the test ran to completion.
